refactor(bookmarks): group bookmark editor components
