+++ src/features/bookmarks/components/bookmark-editor/bookmark-tag-field/index.stories.tsx
@@ -3,10 +3,10 @@ import { expect, fn, userEvent, within } from 'storybook/test'
 import { styled } from 'styled-system/jsx'
 import * as v from 'valibot'
 
-import { err, ok } from '../../../../shared/domain/result'
-import preview from '../../../../storybook/preview'
-import { tagIdSchema, tagNameSchema } from '../../../tags/domain/tag-values'
-import type { TagId } from '../../../tags/domain/tag-values'
+import { err, ok } from '../../../../../shared/domain/result'
+import preview from '../../../../../storybook/preview'
+import { tagIdSchema, tagNameSchema } from '../../../../tags/domain/tag-values'
+import type { TagId } from '../../../../tags/domain/tag-values'
 import { BookmarkTagField } from './index'
 import type { CreateTag, SelectableTag } from './index'
